requirements.txt
google-cloud-texttospeech==0.1.0
google-cloud-speech==0.33.0
google-cloud-vision==0.30.1
